consolidate db files
