+++ src/lib/database/setup.ts
@@ -5,7 +5,7 @@ const setup = async (pool: Pool): Promise<void> => {
     
     const sql = await fs
         .readFile(
-            `${__dirname}/../../sql/setup.sql`,
+            `${__dirname}/../../../sql/setup.sql`,
             { encoding: 'utf-8' }
         );
